No SQL Keywords Columns
The new version of H2 does not like it if SQL keywords are used as
column names, causing several tests to fail.

Since we do not actually care for those names in the test specific
tables, this just changes them to `a` and `b`.  In the non-test cases,
we try escaping instead.